Immediately compile backwards graph in AOTAutograd
Previously, we made backwards graph compilation lazy to avoid paying
for compilation if the user didn't actually end up using the backwards
graph.  This was useful in the old days when a lot of things in Inductor
didn't work and we could bypass errors this way.

However, this has a bad implication for dynamic shapes: the backwards
graph compilation can trigger extra guards, which are too late to
install in the Dynamo context if we wait until backwards is being run.
So in this PR I move us back to compiling backwards graph immediately.
This should also make it easier to predict when compilation occurs,
since compilation now all happens up front during forwards.

Signed-off-by: Edward Z. Yang <ezyangmeta.com>

ghstack-source-id: 47eb2c43319cd4cc968d17a4cb81e4df5a723f8e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/104971