Transmute refined SymInt into int
Previously, x.size(0) could return a SymInt, even when the internal
sympy expression was actually already constant (e.g., due to an
introduced guard.)  We now allow to query the Python object with
maybe_as_int which allows us to transmute these objects back to
int when possible.

It is still possible to end up with a constant SymInt even after this
change, e.g., if you get out a SymInt and while holding onto it
specialize it, but casual users are more likely to get ints when they
want to.

Signed-off-by: Edward Z. Yang <ezyangmeta.com>

ghstack-source-id: d785ec119f8e516c239740838fe8c6b8bf6d16fc
Pull Request resolved: https://github.com/pytorch/pytorch/pull/104828